+++ include/REL/REL.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "REL/Common.h"
+
+#include "REL/ID.h"
+#include "REL/Module.h"
+#include "REL/Offset.h"
+#include "REL/Pattern.h"
+#include "REL/Relocation.h"
+#include "REL/Version.h"
